Fix tests including mocked
